+++ NEWS.md
@@ -1,3 +1,5 @@
+# textrecipes (development version)
+
 # textrecipes 0.0.1
  
 First CRAN version