+++ NEWS.md
@@ -1,5 +1,7 @@
 # textrecipes (development version)
 
+* Custom stemming function can now be used in step_stem using the custom_stemmer argument.
+
 # textrecipes 0.0.1
  
 First CRAN version