+++ NEWS.md
@@ -1,6 +1,7 @@
 # textrecipes (development version)
 
 * Custom stemming function can now be used in step_stem using the custom_stemmer argument.
+* `step_textfeatures()` have been added, allows for multiple numerical features to be pulled from text.
 
 # textrecipes 0.0.1