+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Custom stemming function can now be used in step_stem using the custom_stemmer argument.
 * `step_textfeatures()` have been added, allows for multiple numerical features to be pulled from text.
+* `step_sequence_onehot()` have been added, allows for one hot encoding of sequences of fixed width.
 
 # textrecipes 0.0.1