+++ NEWS.md
@@ -3,6 +3,7 @@
 * Custom stemming function can now be used in step_stem using the custom_stemmer argument.
 * `step_textfeatures()` have been added, allows for multiple numerical features to be pulled from text.
 * `step_sequence_onehot()` have been added, allows for one hot encoding of sequences of fixed width.
+* `step_tokenmerge()` have been added, combines multiple list columns into one list-columns.
 
 # textrecipes 0.0.1