+++ NEWS.md
@@ -1,4 +1,4 @@
-# textrecipes (development version)
+# textrecipes 0.0.2
 
 * Custom stemming function can now be used in step_stem using the custom_stemmer argument.
 * `step_textfeatures()` have been added, allows for multiple numerical features to be pulled from text.